Move remove loser and rename filehandler to helper-functions
